First pass of AI-generated code

+++ backend2/src/main/java/com/lastcallsw/trackeats/repositories/NutritionRepository.java
@@ -3,7 +3,7 @@ package com.lastcallsw.trackeats.repositories;
 import org.springframework.data.repository.CrudRepository;
 import com.lastcallsw.trackeats.entities.Nutrition;
 
-public interface NutritionRepository extends CrudRepository<Nutrition, Long> {
+public interface NutritionRepository extends CrudRepository<Nutrition, Integer> {
     // This class is intentionally left empty. It extends CrudRepository to provide CRUD operations for Nutrition.
     // You can add custom query methods here if needed.
